Improved auth and routing for better structure and navigation.
