+++ frontend/src/utils/PrivateRoute.jsx
@@ -1,0 +1,9 @@
+import { Navigate, Outlet, useLocation } from 'react-router-dom';
+
+const PrivateRoute = ({ isAuthenticated }) => {
+  const location = useLocation();
+  return isAuthenticated ? <Outlet /> :  <Navigate to="/auth" replace state={{ from: location.pathname }} 
+/>;
+};
+
+export default PrivateRoute;
